azure-functions/requirements.txt
azure-functions
sendgrid>=6.9.0

